Add ostream for node, create fresh snode in create_anode
if we don't create fresh snode in create_anode function, we need to deal
with the problem that the old snode (cur->values[pos])'s txn filed will
be assigned the new created anode value in that 2 cas operations. But
you want the snode (which under the newly created anode)'s txn field is
notxn.